Transpile to ES5 and deploy that instead of the original ES6
It is easier to use for downstream projects if this module is
transpiled to ES5 syntax. E.g. in MetaMask, we couldn't figure out how
to add this library as a dependency without it being ES5.

When using the transpiled code in MyEtherWallet, everything works, but
there is this new warning:

```

 warning  in ./node_modules/bitbox02-api/lib/bitbox02-api-go.js

Critical dependency: require function is used in a way in which dependencies cannot be statically extracted
```

This file is copied as is without changes, so it is not clear why this
warning would appear now.

In tests, the target is node so jest works, according to https://jestjs.io/docs/getting-started#using-babel