Remove additional offset of 27
Since
https://github.com/digitalbitbox/bitbox02-api-go/commit/971fa13fa6d4f3cf78df85ddfe6fcad8a4b541c2,
the offset of 27 is added already in the Go library.

The update was pulled in in c4379142f5c596327eab20e9bc9f3b53e6c70b46.

Adding the offset again is incorrect.

`parseInt(_, 16)` was just wrong - it is not a hex string but a
bytestring. It worked by accident since the recID was in the range
0..3, which is parsed the same way in decimal and in hex.