drop BitBox02 response on a closed connection
When closing a WebHID connection, the inputreport callback can apparently still
be called for a closed device when a *new* device is opened. I assume
WebHID dispatches the event based on some sort of device ID which
remains the same, though the old `device` connection was in fact
closed :shrug:

We remove the inputreport event listener after close to solve this,
and perform a validity check for more robustness.