add option to force using the BitBoxBridge over WebHID
Doing WebHID first does not work in Chrome extension background
scripts. There, `navigator.hid` is defined but `requestDevice()`
always returns an empty list. There seems to be no way to figure out
if `navigator.hid` will work before actually trying to use it.

We are likely running into this check in the extension:

https://github.com/chromium/chromium/blob/master/content/browser/hid/hid_service.cc#L144

With this, Chrome extensions can use the forceBridge hint to skip
WebHID.