convert bitbox02-api-js to use a module export
The original change is in main.go, changing the export of the bitbox02
var to be in module.exports instead of a global. The rest of the diff
is follow up work to make the demo work again by converting it to a
webpack project.

The new webpack based sandbox was created by following
https://webpack.js.org/guides/getting-started/ and then copying over
the index.html and assets from the existing demo/.

This change is needed so downstream webpack projects can properly do
code splitting when lazily loading the bitbox02-api package.

The change is internal, so existing downstream projects should be able
to upgrade without an issue.