api: allow a zero TTL value in Cardano transactions
See also: https://github.com/digitalbitbox/bitbox02-firmware/pull/844

Before, ttl="0" was interpreted as "not provided", which means that it
is not serialized and signed. Now we allow ttl="0", and also allow
ttl=undefined to indicate that it is missing.

According to the spec, zero is a valid value for the TTL. It means
however that the transaction can never be mined, as the current slot
will always be bigger than 0.

According to the Adalite team, this type of transaction is used in
meta protocols:
https://github.com/vacuumlabs/adalite/pull/1212#discussion_r755973295